src/serving/requirements.txt
functions-framework
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-pubsub
google-cloud-firestore
google-genai
tenacity
pandas
pandas-gbq
db-dtypes
requests
matplotlib
mplfinance
tweepy
